requirements.txt
requests
pandas
setuptools
wheel
numpy==1.23.5
sqlalchemy
pydantic
mysqlclient
tenacity
fastapi
uvicorn
yfinance
colorlog
finnhub-python
dotenv
pymysql